Refactor Subscriptions model so channel_id is the new key
By setting channel_id as the new key, a single person no longer has
multiple subscription structs in the system, but a single one with all
the lines in it. Some pitfalls to the current model is that everytime a
new subscription is saved, tfl lines get overwritten, so the new
structure should contain old and new.