Centralize all commands to single endpoint
As I looked further into how to integrate with Telegram, I figured that
unlike Slack, it doesn't allow you to use a different endpoint for
different slash commands: it simply allows you to set a single webhook
for all interaction with the bot.

With that in mind, I've gone one step back with the Slack integration
and centralized all the logic in a single endpoint with a specific
handler for that endpoint. On the other hand, the different commands
logic has been segregated into it's own module so when we create a new
endpoint for telegram, that can be reused.